requirements.txt
blinker==1.7.0
click==8.1.7
Empty==0.4.4
grpcio==1.59.3
grpcio-tools==1.59.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
protobuf==4.25.1
setuptools==69.0.2
six==1.16.0
Werkzeug==3.0.1
